Refactor StringTokenScanner to its file.
